refactor: renamed to internal
